컨트롤러의 작업을 서비스에게 위임하라
- 기존에는 컨트롤러가 서비스 로직을 모두 포함하고 있었는데, 이는 응집도가 낮아진다는 문제점이 있습니다. 그래서 컨트롤러가 하던 일들을 서비스 객체로 분리하였습니다.
