+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,106 +1,48 @@
 package com.codesoom.assignment.controllers;
 
 import com.codesoom.assignment.models.Task;
+import com.codesoom.assignment.services.TaskService;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
-import java.util.List;
 
 @RestController
 @RequestMapping("/tasks")
 public class TaskController {
-    private List<Task> tasks = new ArrayList<>();
-    private Long newId = 0L;
+    private final TaskService taskService;
+    public TaskController(TaskService taskService) {
+        this.taskService = taskService;
+    }
 
     @GetMapping
-    public List<Task> getList() {
-        return tasks;
+    public Object getList() {
+        return taskService.handleList();
     }
 
     @GetMapping("{id}")
     public Object getItem(@PathVariable("id") Long id, HttpServletResponse response) {
-        Task task = findTask(id);
-
-        if (task == null) {
-            response.setStatus(404);
-            return "";
-        }
-
-        return task;
+        return taskService.handleItem(id, response);
     }
 
     @PostMapping
     public Object create(@RequestBody Task task, HttpServletResponse response) {
-        if(task.getTitle().isBlank()) {
-            response.setStatus(400);
-            return "";
-        }
-        else {
-            task.setId(generateId());
-            tasks.add(task);
-            response.setStatus(201);
-            return task;
-        }
+        return taskService.handleCreate(task, response);
     }
 
     @PutMapping("/{id}")
     public Object update(@PathVariable("id") Long id,
                          @RequestBody Task source, HttpServletResponse response) {
-        return handleUpdate(id, source, response);
+        return taskService.handleUpdate(id, source, response);
     }
 
     @PatchMapping("/{id}")
     public Object patch(@PathVariable("id") Long id,
                         @RequestBody Task source, HttpServletResponse response) {
-        return handleUpdate(id, source, response);
+        return taskService.handleUpdate(id, source, response);
     }
 
     @DeleteMapping(value = {"", "/", "/{id}"})
     public String delete(@PathVariable(required = false) Long id, HttpServletResponse response) {
-        if(id == null) {
-            response.setStatus(404);
-            return "";
-        }
-
-        Task task = findTask(id);
-
-        if(task == null) {
-            response.setStatus(404);
-            return "";
-        }
-
-        response.setStatus(204);
-        tasks.remove(task);
-        return "";
-    }
-
-    private Task findTask(Long id) {
-        return tasks.stream()
-                .filter(task -> task.getId().equals(id))
-                .findFirst()
-                .orElse(null);
-    }
-
-    private Long generateId() {
-        newId += 1;
-        return newId;
-    }
-
-    private Object handleUpdate(Long id, Task source, HttpServletResponse response) {
-        if(source.getTitle().isBlank()) {
-            response.setStatus(400);
-            return "";
-        }
-
-        Task task = findTask(id);
-
-        if (task == null) {
-            response.setStatus(404);
-            return "";
-        }
-
-        task.setTitle(source.getTitle());
-        return task;
+        return taskService.handleDelete(id, response);
     }
 }
